Don't send breakpoints from lldb_command for symbols without sources
lib/lldb_commands.py:_GetBreaks iterates over all found breakpoints
and checks the filename of it. However, lldb will set breakpoints in
targets with symbol information but no sources and thus the code
tries to call functions on `None`.

Break out of the inner loop if there is no filename corresponding to
the breakpoint