Add a check to remove tab-persistant windows and a split/vertical option
Plugins such as nerdtree-tabs persist windows automatically on every
pane. So the call to `tabnew | split` would be interfered with via
this automatic second window. So upon opening nvim-gdb we should
clear the fixed settings for these windows and close all others.

Also, add an option to have either horizontal or vertical splits for
the two default windows.