Enter window and :e file instead of autocommands
The `doautoa` command runs autocommands on all buffers available. For
small numbers of buffers this isn't a problem. But as you debug a
larger project things get REALLY slow as it runs autocommands on every
file that is open.

There doesn't seem to be an api to say "run this autocommand on that
buffer/window/tab". `doauto` only does it on the currently selected
buffer. Instead we have to manually enter that window/buffer and then
rely on manually opening the file.