revert back to the old lldb implementation
You can use `debugger.GetID()` to uniquely identify a debugger and
`lldb.SBDebugger_FindDebuggerWithID(id)` to get it later in a different
context. This allows us to restore the old implementation.

This is ideal because the current lldb proxy doesn't correctly handle
input all the time. Isolating the exact issue is difficult, but many
keys entries can trigger an erasure of the entire line as well as
frequent clobbering of history.